requirements.txt
python-decouple
djangorestframework
django-cors-headers
djangorestframework-simplejwt